changed project structure: divided stores and functions
