--- src/lib/data/userInitialsStore.js
@@ -1,5 +1,0 @@
-import { writable } from 'svelte/store';
-
-const userInitials = writable('');
-
-export { userInitials };
