+++ src/lib/stores/packages.js
@@ -3,8 +3,6 @@ import { writable, derived } from 'svelte/store';
 import { db } from '../fbconfig';
 
 export const allPackages = writable([]);
-export const allUsers = writable([]);
-export const allKeys = writable([]);
 
 export async function initPackages() {
     const colq = collection(db, 'packages');
@@ -28,26 +26,6 @@ export const incomingPackages = derived([allPackages], ([$packages]) => {
 export const activePackages = derived([allPackages], ([$packages]) => {
     return $packages.filter((pack) => pack.accepted !== '');
 });
-
-export async function initUsers() {
-    const colq = collection(db, 'users');
-
-    const snapshot = await getDocs(colq);
-
-    const loadedUsers = snapshot.docs.map((doc) => ({ id: doc.id, ...doc.data() }));
-
-    allUsers.set(loadedUsers);
-}
-
-export async function initKeys() {
-    const colq = collection(db, 'keys');
-
-    const snapshot = await getDocs(colq);
-
-    const loadedKeys = snapshot.docs.map((doc) => ({ id: doc.id, ...doc.data() }));
-
-    allKeys.set(loadedKeys);
-}
 
 export async function activatePackage(id) {
     const timestamp = new Date().toISOString();
