+++ src/lib/stores/users.js
@@ -1,0 +1,16 @@
+import { writable } from "svelte/store";
+import { db } from '$lib/fbconfig.js'
+import { collection, getDocs } from "firebase/firestore";
+
+export const allUsers = writable([]);
+export const userInitials = writable([]);
+
+export async function initUsers() {
+    const colq = collection(db, 'users');
+
+    const snapshot = await getDocs(colq);
+
+    const loadedUsers = snapshot.docs.map((doc) => ({ id: doc.id, ...doc.data() }));
+
+    allUsers.set(loadedUsers);
+}
